// src/api/schema.ts
export interface components {
    schemas: {
        /**
         * @description Статус задачи
         * @enum {string}
         */
        TaskStatus: "unassigned" | "assigned" | "in_progress" | "done";
// ...
        ProjectTaskResponse: {
            /**
             * @description Уникальный идентификатор задачи
             * @example 7
             */
            id: number;
            /**
             * @description ID проекта
             * @example 10
             */
            project_id: number;
            /**
             * @description ID спринта
             * @example 2
             */
            sprint_id?: number | null;
            /**
             * @description Название задачи
             * @example Реализовать Swagger API
             */
            title: string;
            /** @description Описание задачи */
            description?: string | null;
            status: components["schemas"]["TaskStatus"];
            /** @enum {integer} */
            priority: 1 | 2 | 3 | 4 | 5;
            /**
             * Format: date-time
             * @description Дата создания задачи
             */
            created_at: string;
        };
// ...
    };
// ...
}